Link first activity to second
Now if you press the OK button on the first activity it will open the
second activity via an intent, and that activity's onCreate method has
been updated to add text to the textView. The second activity's layout
has also been changed from relative to linear, because it looks better
that way.