Add animal model for storing holding animal knowledge
Each animal contains a name, a direct question to identify it, a boolean
for whether the answer should be yes or no for that question and the
next animal (for when this was not the right one).

This class should eventually be JSONable so that it can be sent to other
classes via intent message, and Serializeable so that it can be stored
in a file or database, whichever will be chosen later on.