Change data structure of animal from linear to tree
There are cases when both the Yes and the No answer for a question will
*not* lead to an animal, therefore the data structure should be a binary
tree instead of a linear structure. Possibly it also means that "Animal"
is not really a suiting name for this class. Since the addNext method
provided the redundant function of setting a public member, it has been
removed. The functions for parsing the user choice have also been
updated to handle this new structure.