Set appropriate SDK minimum and target in manifest
Both target SDK version and minimum SDK version were set to 8 under the
assumption that that was how to make the app compatible with older
versions. The app has been tested to work (so far) as low down as SDK
version 4 (because version 3 doesn't support the onClick listener), so
the application now runs from SDK version 4 to 18 (currently the latest
release of Android) with these attributes set appropriately.