Add prevAnimal to track Animal parents in the binary tree
After adding a new Animal to the end of the tree, we need to save the
tree to Game.firstAnimal. To do that, we need to find the first Animal
in the tree. So, to traverse it backwards I've added a prevAnimal field
which tracks the parents. Now, we can use a while loop to find the last
animal with null for a parent (that's the first animal) and save it to
Game.firstAnimal.