Bump min and target SDK versions
As part of of modernising this project in 2017 I've bumped up the
minimum SDK version from 4 to 9 since statistics show that 0% of users
are still using phones with as low a version as 8.  I've also bumped the
target SDK version to 25, the last stable version currently.

In summary:

 - minSdkVersion     4 -> 9
 - targetSdkVersion 21 -> 25

These values now come from build.gradle, which overrides those in the
AndroidManifest.xml file, so I've removed them from there to avoid
confusion.  Fixes #9.