Add configuration file for Danger meta checker
So far it checks the following things:

 - PR log doesn't contain merges
 - git commit style
 - Androind linting
 - files contain no TODOs or FIXMEs

It also contains a trigger for Coveralls code coverage measurement
because I didn't want to make a separate ruby file for that.

Unfortunately the plugins providing the above functionality required me
to add a Gemfile for these dependencies but at least it's minimal and I
don't have to bother installing ruby because Travis already provides it.