Refactor paddings in the activity xmls
There is two different margin added to the dimens.xml,
which have the same value and used as paddings in the
activity xmls. A global android:padding can be used for
setting all the padding for one ui element, so the
refactor goes to this changes.

Also the project has 2 dimens.xml files for tablets,
but as this project is still in a very early stage,
I think it's better to just remove those files.