Delete broken Danger android_lint plugin
I installed this plugin on a whim without anything to try it on but it
errors out, looking for a lint reports file called:

    lint-result.xml

but the file generated by the build is

    lint-results.xml

with an S, so it doesn't find it.  The plugin doesn't accept overriding
the file name as a parameter, so I'm removing it, at least for now.