Check user provides input before continuing learning
When learning an animal, the fields for `name` and `question` are
required, so it doesn't make sense to let the user get to the next
activity if they press the button without entering any text.

I've added a check to make sure those user input fields contain more
than just whitespace before sending the user on to the next activity.

Includes translations for:

  * Afrikaans
  * English
  * Hungarian

Fixes #5.