Upgrade app theme to Material Design with support library
This fixes #10, that the app looked ugly and outdated because it was
using the theme provided with Android version 4.  It also fixes #42,
that the text field error popup text was white-on-white, and thus
unreadable, which was basically also an issue with the theme.

The solution here was simply to use the design support library to access
the Material Design theme with backwards compatibility and set a
Material Design theme (I chose "Light" with no Action Bar) as the app
theme.  Everything looks fine now.