Persist learnt animals to internal file
Currently learnt animals would be forgotten when you restart the app
because they were only stored in-memory in the game state.  This commit
solves that by implementing reading and writing to Android internal
storage.

It serialises the learnt animals from the game state to a file after the
last learning step.  If this file exists when the application starts it
will then deserialise the animal objects from the file and put them into
the game state for use.  If the file does not exist it will populate the
game state with the 2 example animals we used so far.

As far as I can tell, the exceptions that could arise from this file use
are highly unlikely and would not be the user's fault, so they are
simply "WTF" logged.

Related GitHub issues:

- Fixes #4 Forgets animals on restart
- Resolves #13 Store Animals locally
- Unblocks #36 Have a menu with Options, About & Help etc.