Revert the refactor of learnNewAnimal arguments
The previous commit simplified the way the learnNewAnimal method
receives the nextReq value by getting it directly from its class's field
instead of as an argument.  However, the method is only sometimes called
with prevReq (the user's response to the previous question) as an
argument, the rest of the time it's called with choice which is the
result of their current interaction. So it's not as simple as just using
the value we have in prevReq.

I'm reverting that part of the patch here.