Simplify GuessActivity starter calls with defaults
Where GuessActivity is started from outside itself, it's always called
with the same arguments:

    context, firstAnimal, false, false

The context has to be provided by the caller, but the firstAnimal is
stored in the static Game state and the other two are literally false.
I put the last 3 into a convenience method that overloads the existing
start method.  This makes those calls much simpler because now the only
argument they need to give is `this` for context.